Crypto.Hash

Description

Crypto hash main module

Types

class HashAlgorithm a where

Class representing hashing algorithms.

The hash algorithm is built over 3 primitives:

init : create a new context updates : update the context with some strict bytestrings finalize : finalize the context into a digest

Methods

hashInit :: Context a

Initialize a new context for this hash algorithm

hashUpdates :: Context a -> [ByteString] -> Context a

Update the context with a list of strict bytestring, and return a new context with the updates.

hashFinalize :: Context a -> Digest a

Finalize a context and return a digest.

digestFromByteString :: ByteString -> Maybe (Digest a)

Try to convert a binary digest bytestring to a digest.

data Context a

Represent a context for a given hash algorithm.

data Digest a

Represent a digest for a given hash algorithm.

Functions

digestToByteString :: Digest a -> ByteString

Return the binary digest

digestToHexByteString :: Digest a -> ByteString

Return the hexadecimal (base16) bytestring of the digest

hash :: HashAlgorithm a => ByteString -> Digest a

Hash a strict bytestring into a digest.

hashlazy :: HashAlgorithm a => ByteString -> Digest a

Hash a lazy bytestring into a digest.

hashUpdate :: HashAlgorithm a => Context a -> ByteString -> Context a

run hashUpdates on one single bytestring and return the updated context.

hash algorithms

data MD2

data MD4

data MD5

data SHA1

data SHA224

data SHA256

data SHA384

data SHA512

data RIPEMD160

data Tiger

data SHA3_224

data SHA3_256

data SHA3_384

data SHA3_512

data Skein256_224

data Skein256_256

data Skein512_224

data Skein512_256

data Skein512_384

data Skein512_512

data Whirlpool