Crypto.Hash.Types

Description

Crypto hash types definitions

Documentation

class HashAlgorithm a where

Class representing hashing algorithms.

The hash algorithm is built over 3 primitives:

init : create a new hashing context

updates : update the hashing context with some strict bytestrings and return the new context

finalize : finalize the context into a digest

Methods

hashBlockSize :: Context a -> Int

Block size in bytes the hash algorithm operates on

hashInit :: Context a

Initialize a new context for this hash algorithm

hashUpdates :: Context a -> [ByteString] -> Context a

Update the context with a list of strict bytestring, and return a new context with the updates.

hashFinalize :: Context a -> Digest a

Finalize a context and return a digest.

digestFromByteString :: ByteString -> Maybe (Digest a)

Try to convert a binary digest bytestring to a digest.

newtype Context a

Represent a context for a given hash algorithm.

Constructors

Context ByteString

newtype Digest a

Represent a digest for a given hash algorithm.

Constructors

Digest ByteString

deprecated

contextToByteString :: Context a -> ByteString

return the binary bytestring. deprecated use toBytes.

digestToByteString :: Digest a -> ByteString

Deprecated: use toBytes from byteable:Data.Byteable

return the binary bytestring. deprecated use toBytes.