Crypto.Hash

Description

Generalized cryptographic hash interface, that you can use with cryptographic hash algorithm that belong to the HashAlgorithm type class.

import Crypto.Hash

sha1 :: ByteString -> Digest SHA1
sha1 = hash

hexSha3_512 :: ByteString -> String
hexSha3_512 bs = show (hash bs :: Digest SHA3_512)

Types

class HashAlgorithm a where

Class representing hashing algorithms.

The hash algorithm is built over 3 primitives:

init : create a new hashing context

updates : update the hashing context with some strict bytestrings and return the new context

finalize : finalize the context into a digest

Methods

hashBlockSize :: Context a -> Int

Block size in bytes the hash algorithm operates on

hashInit :: Context a

Initialize a new context for this hash algorithm

hashUpdates :: Context a -> [ByteString] -> Context a

Update the context with a list of strict bytestring, and return a new context with the updates.

hashFinalize :: Context a -> Digest a

Finalize a context and return a digest.

digestFromByteString :: ByteString -> Maybe (Digest a)

Try to convert a binary digest bytestring to a digest.

type HashFunctionBS a = ByteString -> Digest a

Alias to a single pass hash function that operate on a strict bytestring

type HashFunctionLBS a = ByteString -> Digest a

Alias to a single pass hash function that operate on a lazy bytestring

data Context a

Represent a context for a given hash algorithm.

data Digest a

Represent a digest for a given hash algorithm.

Functions

digestToByteString :: Digest a -> ByteString

Deprecated: use toBytes from byteable:Data.Byteable

return the binary bytestring. deprecated use toBytes.

digestToHexByteString :: Digest a -> ByteString

Return the hexadecimal (base16) bytestring of the digest

hash :: HashAlgorithm a => ByteString -> Digest a

Hash a strict bytestring into a digest.

hashlazy :: HashAlgorithm a => ByteString -> Digest a

Hash a lazy bytestring into a digest.

hashUpdate :: HashAlgorithm a => Context a -> ByteString -> Context a

run hashUpdates on one single bytestring and return the updated context.

hashInitAlg :: HashAlgorithm alg => alg -> Context alg

Initialize a new context for a specified hash algorithm

hash algorithms

data MD2

MD2 cryptographic hash algorithm

Constructors

MD2

data MD4

MD4 cryptographic hash algorithm

Constructors

MD4

data MD5

MD5 cryptographic hash algorithm

Constructors

MD5

data SHA1

SHA1 cryptographic hash algorithm

Constructors

SHA1

data SHA224

SHA224 cryptographic hash algorithm

Constructors

SHA224

data SHA256

SHA256 cryptographic hash algorithm

Constructors

SHA256

data SHA384

SHA384 cryptographic hash algorithm

Constructors

SHA384

data SHA512

SHA512 cryptographic hash algorithm

Constructors

SHA512

data RIPEMD160

RIPEMD160 cryptographic hash algorithm

Constructors

RIPEMD160

data Tiger

Tiger cryptographic hash algorithm

Constructors

Tiger

data SHA3_224

SHA3 (224 bits) cryptographic hash algorithm

Constructors

SHA3_224

data SHA3_256

SHA3 (256 bits) cryptographic hash algorithm

Constructors

SHA3_256

data SHA3_384

SHA3 (384 bits) cryptographic hash algorithm

Constructors

SHA3_384

data SHA3_512

SHA3 (512 bits) cryptographic hash algorithm

Constructors

SHA3_512

data Skein256_224

Skein256 (224 bits) cryptographic hash algorithm

Constructors

Skein256_224

data Skein256_256

Skein256 (256 bits) cryptographic hash algorithm

Constructors

Skein256_256

data Skein512_224

Skein512 (224 bits) cryptographic hash algorithm

Constructors

Skein512_224

data Skein512_256

Skein512 (256 bits) cryptographic hash algorithm

Constructors

Skein512_256

data Skein512_384

Skein512 (384 bits) cryptographic hash algorithm

Constructors

Skein512_384

data Skein512_512

Skein512 (512 bits) cryptographic hash algorithm

Constructors

Skein512_512

data Whirlpool

Whirlpool cryptographic hash algorithm

Constructors

Whirlpool

MAC algorithms

newtype HMAC a

Represent an HMAC that is a phantom type with the hash used to produce the mac.

The Eq instance is constant time.

Constructors

HMAC

Fields

hmacGetDigest :: Digest a

hmac

Arguments

:: HashAlgorithm a

=> ByteString
Secret key

-> ByteString
Message to MAC

-> HMAC a

compute a MAC using the supplied hashing function

hmacAlg

Arguments

:: HashAlgorithm a

=> a
the hash algorithm the actual value is unused.

-> ByteString
Secret key

-> ByteString
Message to MAC

-> HMAC a

compute a HMAC using a specified algorithm